move project selctor to releases header
