+++ static/app/views/releases/components/header.tsx
@@ -1,5 +1,13 @@
+import styled from '@emotion/styled';
+
+import {Container} from '@sentry/scraps/layout';
+
 import FeedbackButton from 'sentry/components/feedbackButton/feedbackButton';
 import * as Layout from 'sentry/components/layouts/thirds';
+import {DatePageFilter} from 'sentry/components/organizations/datePageFilter';
+import {EnvironmentPageFilter} from 'sentry/components/organizations/environmentPageFilter';
+import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
+import {ProjectPageFilter} from 'sentry/components/organizations/projectPageFilter';
 import {PageHeadingQuestionTooltip} from 'sentry/components/pageHeadingQuestionTooltip';
 import {t} from 'sentry/locale';
 
@@ -27,6 +35,22 @@ export default function Header() {
           }}
         />
       </Layout.HeaderActions>
+      <Container paddingTop="xl">
+        <ReleasesPageFilterBar condensed>
+          <ProjectPageFilter />
+          <EnvironmentPageFilter />
+          <DatePageFilter
+            disallowArbitraryRelativeRanges
+            menuFooterMessage={t(
+              'Changing this date range will recalculate the release metrics. Select a supported date range from the options above.'
+            )}
+          />
+        </ReleasesPageFilterBar>
+      </Container>
     </Layout.Header>
   );
 }
+
+const ReleasesPageFilterBar = styled(PageFilterBar)`
+  grid-column: 1 / -1;
+`;
